+++ src/sprite/SpriteManager.h
@@ -9,12 +9,12 @@
 namespace trippin {
     class SpriteManager {
     public:
-        SpriteManager(SDL_Renderer *renderer, SpriteLoader &spriteLoader, int tickPeriod);
+        SpriteManager(SDL_Renderer *renderer, SpriteLoader &spriteLoader, double tickPeriod);
         const Sprite &get(const std::string &type);
         void setSurfaces(std::unique_ptr<std::unordered_map<std::string, SDL_Surface *>> surfaces);
     private:
         using SpritePtr = std::unique_ptr<Sprite>;
-        const int tickPeriod;
+        const double tickPeriod;
         SDL_Renderer *renderer;
         SpriteLoader &spriteLoader;
         std::unordered_map<std::string, SpritePtr> sprites;
